fix: avoid accidental side effect breakage
Moved the side effect heavy initialization logic that previously
resided in main.py to it's own file that is added at the top.
Should make it less likely to accidentally break things.
